requirements.txt
git+https://github.com/openai/CLIP.git
streamlit==1.26.0
torch
torchvision 
pillow 
matplotlib 
pandas 
transformers
Flask==2.3.2
Werkzeug==2.3.7
sqlite3
sentence-transformers==4.0.1
scikit-learn==1.3.0
numpy==1.24.4
huggingface-hub==0.25.2
plotly==5.15.0
flask_toastr
